a2aNetwork/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "a2a-network"
version = "1.0.0"
description = "A2A Network Infrastructure - Registry, Trust, and SDK Components"
readme = "readme.md"
requires-python = ">=3.8"
license = {text = "MIT License"}
authors = [
  {email = "dev@a2a.network", name = "A2A Network Team"},
]
classifiers=[
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: System :: Distributed Computing",
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "httpx>=0.25.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "redis>=5.0.0",
    "celery>=5.3.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "opentelemetry-instrumentation-fastapi>=0.41b0",
    "opentelemetry-instrumentation-requests>=0.41b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.41b0",
]

[project.urls]
Homepage = "https://github.com/a2a-network/a2a-network"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]
blockchain = [
    "web3>=6.10.0",
    "eth-account>=0.9.0",
    "eth-keys>=0.4.0",
]

[project.scripts]
a2a-registry = "registry.runRegistryServer:main"
a2a-trust = "trustSystem.service:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
exclude = ["node_modules*", "out*", "broadcast*", "cache*", "gen*", "lib*", "*egg-info*", "*pycache*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.json", "*.sql", "*.md"]
